Allow specifying output file in generate.py

+++ generate.py
@@ -34,6 +34,7 @@ def main():
 	script, *files = sys.argv
 	from_branch = None
 	flag_c = False
+#TODO: use some more flexible parser
 	if files and files[0] == '-c':
 		from_branch = 'code'
 		files = files[1:]
@@ -42,6 +43,10 @@ def main():
 	if files and files[0] == '-d':
 		files = files[1:]
 		diff = True
+	output_file = None
+	if files and files[0] == '-o':
+		output_file = files[1]
+		files = files[2:]
 	content_branch = None
 	if not files:
 		files = ['description.txt']
@@ -54,19 +59,22 @@ def main():
 		content.append(escape(get_file(filename, content_branch)))
 	output = '[' + ','.join(content) + ']'
 	
-	to_process = [['index.html', 'output.html']]
-	if len(content) == 1:
-		to_process.append(['quiz.html', 'quiz-output.html'])
+	if output_file is None:
+		to_process = [['index.html', 'output.html']]
+		if len(content) == 1:
+			to_process.append(['quiz.html', 'quiz-output.html'])
+		else:
+			try:
+				os.remove('quiz-output.html')
+			except FileNotFoundError:
+				pass
 	else:
-		try:
-			os.remove('quiz-output.html')
-		except FileNotFoundError:
-			pass
+		to_process = [['index.html', output_file]]
 
 	for input_name, output_name in to_process:
 		content = get_file(input_name, from_branch)
 		with open(output_name, 'w') as output_file:
-			content = content.replace('resources', 'resources_main').replace('hardcoded=undefined', 'hardcoded=' + output);
+			content = content.replace('./resources', 'file:///' + os.path.join(os.getcwd(), 'resources_main')).replace('hardcoded=undefined', 'hardcoded=' + output);
 			print(content, file=output_file)
 	subprocess.run(['rm', '-rf', 'resources_main'])
 	os.mkdir('resources_main')
